--- mod2/src/main/java/FileName.java
@@ -1,3 +1,0 @@
-public class FileName {
-    
-}
