--- mod2/src/main/java/JSONRunner.java
@@ -1,14 +1,0 @@
-import java.net.URISyntaxException;
-import java.util.Scanner;
-
-public class JSONRunner {
-    public static void main(String[] args) throws URISyntaxException {
-        Configuration configuration = new Configuration();
-        configuration.createFilesInJava();
-        configuration.fromJSON();
-        System.out.println( configuration
-                .checkTheFiles(configuration.createFilesInJava(), configuration.fromJSON()));
-
-
-    }
-}
