--- mod2/src/main/resources/test.json
@@ -1,15 +1,0 @@
-{
-  "suffix": "Test",
-  "file": [
-    {
-    "fileName": "1.txt"
-    },
-    {
-      "fileName": "2.txt"
-    },
-    {
-      "fileName": "3.txt"
-    }
-  ]
-
-}
